requirements.txt
pyaudio
openai
pygame
SpeechRecognition
numpy
sounddevice
pyserial
realtimetts[gtts,openai]
ffmpeg
mpv